Create css and classes folder and move related files into their respective folders
