+++ sudoku-solver/src/classes/Cell.js
